backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
greenlet>=3.0.0
asyncpg>=0.30.0
pgvector>=0.3.6
openai>=1.60.0
openai-agents>=0.14.0
pydantic-settings>=2.7.0
httpx>=0.28.0
langfuse>=2.60.0
mcp>=1.4.1
PyJWT[crypto]>=2.10.1
